+++ src/main/java/lotto/dao/LottosDao.java
@@ -2,7 +2,7 @@ package lotto.dao;
 
 import lotto.domain.Lotto;
 import lotto.domain.Lottos;
-import lotto.utils.DBUtils;
+import lotto.db.DBUtils;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
